+++ docs/changelog.rst
@@ -3,7 +3,7 @@ Changelog
 
 next
 ----
-* Added support for Union types.
+* Added support for Union and Literal types.
 * Assume that types annotated as constructible from a single str are their own
   parser.
 * Added support for catching exceptions.